% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_preprocessing.R
\name{authenticate_umls}
\alias{authenticate_umls}
\title{Authenticate with UMLS}
\usage{
authenticate_umls(api_key)
}
\arguments{
\item{api_key}{UMLS API key}
}
\value{
Character string with TGT URL or NULL if authentication fails
}
\description{
This function authenticates with UMLS and returns a TGT URL.
}
\keyword{internal}
