% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_preprocessing.R
\name{create_dummy_dictionary}
\alias{create_dummy_dictionary}
\title{Helper function to create dummy dictionaries}
\usage{
create_dummy_dictionary(dictionary_type)
}
\description{
Helper function to create dummy dictionaries
}
\keyword{internal}
