% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_preprocessing.R
\name{detect_lang}
\alias{detect_lang}
\title{Detect language of text}
\usage{
detect_lang(text, sample_size = 1000)
}
\arguments{
\item{text}{Text string to analyze}

\item{sample_size}{Maximum number of characters to sample for language detection}
}
\value{
Character string containing the ISO 639-1 language code
}
\description{
This function attempts to detect the language of a text string.
It implements a simple n-gram based approach that doesn't require
additional packages.
}
\examples{
\dontrun{
lang <- detect_lang("This is English text")
# Returns "en"
}
}
