% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_preprocessing.R
\docType{data}
\name{.dict_cache_env}
\alias{.dict_cache_env}
\title{Environment to store dictionary cache data}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.dict_cache_env
}
\description{
Environment to store dictionary cache data
}
\keyword{internal}
