% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\docType{data}
\name{.pubmed_cache_env}
\alias{.pubmed_cache_env}
\title{Environment to store PubMed cache data}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.pubmed_cache_env
}
\description{
Environment to store PubMed cache data
}
\keyword{internal}
