% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_preprocessing.R
\name{extract_mesh_from_text}
\alias{extract_mesh_from_text}
\title{Extract MeSH terms from text format instead of XML}
\usage{
extract_mesh_from_text(mesh_text, dictionary_type)
}
\arguments{
\item{mesh_text}{Text containing MeSH data in non-XML format}

\item{dictionary_type}{Type of dictionary}
}
\value{
A data frame with MeSH terms
}
\description{
Extract MeSH terms from text format instead of XML
}
\keyword{internal}
