% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{fetch_and_parse_gene}
\alias{fetch_and_parse_gene}
\title{Fetch and parse Gene data}
\usage{
fetch_and_parse_gene(search_result, max_results, throttle_api, retry_api_call)
}
\description{
Fetch and parse Gene data
}
\keyword{internal}
