% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{fetch_and_parse_pmc}
\alias{fetch_and_parse_pmc}
\title{Fetch and parse PMC data}
\usage{
fetch_and_parse_pmc(search_result, max_results, throttle_api, retry_api_call)
}
\description{
Fetch and parse PMC data
}
\keyword{internal}
