% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{fetch_and_parse_protein}
\alias{fetch_and_parse_protein}
\title{Fetch and parse Protein data}
\usage{
fetch_and_parse_protein(
  search_result,
  max_results,
  throttle_api,
  retry_api_call
)
}
\description{
Fetch and parse Protein data
}
\keyword{internal}
