% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{fetch_and_parse_pubmed}
\alias{fetch_and_parse_pubmed}
\title{Fetch and parse PubMed data}
\usage{
fetch_and_parse_pubmed(
  search_result,
  max_results,
  throttle_api,
  retry_api_call
)
}
\description{
Fetch and parse PubMed data
}
\keyword{internal}
