% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_similarity.R
\name{find_similar_docs}
\alias{find_similar_docs}
\title{Find similar documents for a given document}
\usage{
find_similar_docs(text_data, doc_id, text_column = "abstract", n_similar = 5)
}
\arguments{
\item{text_data}{A data frame containing text data.}

\item{doc_id}{ID of the document to find similar documents for.}

\item{text_column}{Name of the column containing text to analyze.}

\item{n_similar}{Number of similar documents to return.}
}
\value{
A data frame with similar documents and their similarity scores.
}
\description{
This function finds documents similar to a given document based on TF-IDF and
cosine similarity.
}
\examples{
\dontrun{
similar_docs <- find_similar_docs(article_data, doc_id = 1,
                                     text_column = "abstract", n_similar = 5)
}
}
