% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_preprocessing.R
\name{get_dict_cache}
\alias{get_dict_cache}
\title{Get dictionary cache environment}
\usage{
get_dict_cache()
}
\value{
The environment containing cached dictionary data
}
\description{
Get dictionary cache environment
}
