% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{get_pubmed_cache}
\alias{get_pubmed_cache}
\title{Get the pubmed cache environment}
\usage{
get_pubmed_cache()
}
\value{
An environment containing cached PubMed data
}
\description{
Get the pubmed cache environment
}
\keyword{internal}
