% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_preprocessing.R
\name{get_service_ticket}
\alias{get_service_ticket}
\title{Get a service ticket from a TGT URL}
\usage{
get_service_ticket(tgt_url)
}
\arguments{
\item{tgt_url}{Ticket Granting Ticket URL}
}
\value{
Character string with service ticket or NULL if it fails
}
\description{
This function gets a service ticket for a specific service using the TGT URL.
}
\keyword{internal}
