% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_term_vars}
\alias{get_term_vars}
\title{Extract term variations from text corpus}
\usage{
get_term_vars(articles, primary_term, text_col = "abstract")
}
\arguments{
\item{articles}{A data frame containing article data with text columns}

\item{primary_term}{The primary term to find variations of}

\item{text_col}{Name of the column containing the text to search}
}
\value{
A character vector of unique term variations, sorted by length
}
\description{
This function identifies variations of a primary term within a corpus of articles.
}
