% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_preprocessing.R
\name{get_umls_semantic_types}
\alias{get_umls_semantic_types}
\title{Get UMLS semantic types for a given dictionary type}
\usage{
get_umls_semantic_types(dictionary_type)
}
\arguments{
\item{dictionary_type}{The type of dictionary to get semantic types for}
}
\value{
Vector of UMLS semantic type identifiers
}
\description{
Helper function to map dictionary types to UMLS semantic type identifiers
}
\keyword{internal}
