% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_model.R
\name{is_valid_biomedical_entity}
\alias{is_valid_biomedical_entity}
\title{Determine if a term is likely a specific biomedical entity with improved accuracy}
\usage{
is_valid_biomedical_entity(term, claimed_type = NULL)
}
\arguments{
\item{term}{Character string, the term to check}

\item{claimed_type}{Character string, the claimed entity type of the term}
}
\value{
Logical, TRUE if the term is likely a valid biomedical entity, FALSE otherwise
}
\description{
Determine if a term is likely a specific biomedical entity with improved accuracy
}
