% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{null_coalesce}
\alias{null_coalesce}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{An object to test if NULL or empty}

\item{y}{An object to return if x is NULL or empty}
}
\value{
Returns x if x is not NULL, not empty, or not a missing XML node, otherwise returns y.
}
\description{
Returns the first argument if it is not NULL or empty, otherwise returns the second argument.
}
\examples{
NULL \%||\% "default"  # returns "default"
"value" \%||\% "default"  # returns "value"
}
\keyword{internal}
