% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_preprocessing.R
\name{process_mesh_chunks}
\alias{process_mesh_chunks}
\title{Process MeSH data in chunks to avoid memory issues}
\usage{
process_mesh_chunks(mesh_records, dictionary_type)
}
\arguments{
\item{mesh_records}{Large MeSH records data}

\item{dictionary_type}{Type of dictionary}
}
\value{
A data frame with MeSH terms
}
\description{
Process MeSH data in chunks to avoid memory issues
}
\keyword{internal}
