% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_mesh}
\alias{query_mesh}
\title{Query for MeSH terms using E-utilities}
\usage{
query_mesh(term, api_key = NULL)
}
\arguments{
\item{term}{Character string, the term to query.}

\item{api_key}{Character string. NCBI API key (optional).}
}
\value{
A data frame with MeSH information for the term.
}
\description{
Query for MeSH terms using E-utilities
}
