% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_umls}
\alias{query_umls}
\title{Query UMLS for term information}
\usage{
query_umls(term, api_key, version = "current")
}
\arguments{
\item{term}{Character string, the term to query.}

\item{api_key}{Character string. UMLS API key.}

\item{version}{Character string. UMLS version to use.}
}
\value{
A data frame with UMLS information for the term.
}
\description{
Query UMLS for term information
}
