% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_preprocessing.R
\name{segment_sentences}
\alias{segment_sentences}
\title{Perform sentence segmentation on text}
\usage{
segment_sentences(text)
}
\arguments{
\item{text}{Character vector of texts to process}
}
\value{
A list where each element contains a character vector of sentences
}
\description{
This function splits text into sentences.
}
\examples{
\dontrun{
sentences <- segment_sentences(abstracts)
}
}
